+++ CHANGELOG.md
@@ -4,6 +4,21 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
+## Version 0.6.5 - Released 2020-08-26 
+
+### Added
+* Add `to_boxes` to MultiPolygon
+
+### Changed
+* More methods in the Mask object should work without the c-extensions being built. 
+* The `Mask.to_coco` method now returns a format based on the native encoding.
+* Support for the new is preferred "ltrb" format over "tlbr" which will eventually be deprecated.
+* No longer publishing wheels / CI testing for Python 2.7
+
+### Fixed
+* Fix bug with check for turbojpeg in imread.
+* `subpixel_slice` now works with non-integer slices
+
 ## Version 0.6.4 - Released 2020-07-08 
 
 ### Added
@@ -215,6 +230,3 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
     - `im_cv2`
     - `im_alphablend`
 
-## Version 0.6.3 - Unreleased
-
-## Version 0.6.4 - Unreleased